+++ core/design/src/androidTest/kotlin/edu/stanford/spezi/core/design/personalInfo/composables/NameFieldsTestComposable.kt
@@ -8,20 +8,20 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import edu.stanford.spezi.core.design.component.StringResource
 import edu.stanford.spezi.core.design.views.personalInfo.PersonNameComponents
-import edu.stanford.spezi.core.design.validation.personalInfo.fields.NameFieldRow
+import edu.stanford.spezi.core.design.views.personalInfo.fields.NameFieldRow
 
 @Composable
 fun NameFieldsTestComposable() {
-    val name = remember { mutableStateOf(edu.stanford.spezi.core.design.views.personalInfo.PersonNameComponents()) }
+    val name = remember { mutableStateOf(PersonNameComponents()) }
 
     Column {
-        NameFieldRow(StringResource("First Name"), name, edu.stanford.spezi.core.design.views.personalInfo.PersonNameComponents::givenName) {
+        NameFieldRow(StringResource("First Name"), name, PersonNameComponents::givenName) {
             Text("enter your first name")
         }
 
         HorizontalDivider()
 
-        NameFieldRow(StringResource("Last Name"), name, edu.stanford.spezi.core.design.views.personalInfo.PersonNameComponents::familyName) {
+        NameFieldRow(StringResource("Last Name"), name, PersonNameComponents::familyName) {
             Text("enter your last name")
         }
     }
