+++ core/design/src/androidTest/kotlin/edu/stanford/spezi/core/design/views/composables/SuspendButtonTestComposable.kt
@@ -9,9 +9,9 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import edu.stanford.spezi.core.design.component.StringResource
-import edu.stanford.spezi.core.design.validation.views.model.ViewState
-import edu.stanford.spezi.core.design.validation.views.viewModifier.viewState.ViewStateAlert
-import edu.stanford.spezi.core.design.validation.views.views.button.SuspendButton
+import edu.stanford.spezi.core.design.views.views.model.ViewState
+import edu.stanford.spezi.core.design.views.views.viewModifier.viewState.ViewStateAlert
+import edu.stanford.spezi.core.design.views.views.views.button.SuspendButton
 import kotlinx.coroutines.delay
 import kotlin.time.Duration.Companion.milliseconds
