+++ core/design/src/main/kotlin/edu/stanford/spezi/core/design/views/validation/configuration/ValidationEngine.kt
@@ -1,6 +1,6 @@
 package edu.stanford.spezi.core.design.views.validation.configuration
 
 import androidx.compose.runtime.compositionLocalOf
-import edu.stanford.spezi.core.design.validation.validation.ValidationEngine
+import edu.stanford.spezi.core.design.views.validation.ValidationEngine
 
 val LocalValidationEngine = compositionLocalOf<ValidationEngine?> { null }
