+++ core/design/src/main/kotlin/edu/stanford/spezi/core/design/views/validation/configuration/ValidationEngineConfiguration.kt
@@ -1,8 +1,8 @@
 package edu.stanford.spezi.core.design.views.validation.configuration
 
 import androidx.compose.runtime.compositionLocalOf
-import edu.stanford.spezi.core.design.validation.validation.ValidationEngine
-import edu.stanford.spezi.core.design.validation.validation.ValidationEngineConfiguration
+import edu.stanford.spezi.core.design.views.validation.ValidationEngine
+import edu.stanford.spezi.core.design.views.validation.ValidationEngineConfiguration
 
 val LocalValidationEngineConfiguration = compositionLocalOf {
     ValidationEngineConfiguration.noneOf(ValidationEngine.ConfigurationOption::class.java)
