+++ core/design/src/main/kotlin/edu/stanford/spezi/core/design/views/views/viewModifier/viewState/ViewStateMapper.kt
@@ -3,8 +3,8 @@ package edu.stanford.spezi.core.design.views.views.viewModifier.viewState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.MutableState
-import edu.stanford.spezi.core.design.validation.views.model.OperationState
-import edu.stanford.spezi.core.design.validation.views.model.ViewState
+import edu.stanford.spezi.core.design.views.views.model.OperationState
+import edu.stanford.spezi.core.design.views.views.model.ViewState
 
 @Composable
 fun <State : OperationState> MapOperationStateToViewState(state: State, viewState: MutableState<ViewState>) {
